QuickTimeStreamingServer-452

+++ QuickTimeStreamingServer/APIModules/QTSSReflectorModule/RelaySDPSourceInfo.cpp
@@ -1,23 +1,24 @@
 /*
  *
  * @APPLE_LICENSE_HEADER_START@
- *
- * Copyright (c) 1999-2001 Apple Computer, Inc.  All Rights Reserved. The
- * contents of this file constitute Original Code as defined in and are
- * subject to the Apple Public Source License Version 1.2 (the 'License').
- * You may not use this file except in compliance with the License.  Please
- * obtain a copy of the License at http://www.apple.com/publicsource and
- * read it before using this file.
- *
- * This Original Code and all software distributed under the License are
+ * 
+ * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
+ * 
+ * The Original Code and all software distributed under the License are
  * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
- * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.  Please
- * see the License for the specific language governing rights and
+ * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
  * limitations under the License.
- *
- *
+ * 
  * @APPLE_LICENSE_HEADER_END@
  *
  */
@@ -117,7 +118,7 @@ void    RelaySDPSourceInfo::Parse(StrPtrLen* inSDPData)
             fNumOutputs--;// Don't count the ttl as an output!
             
             StringParser ttlParser(&theTtl);
-            theDestTtl = ttlParser.ConsumeInteger(NULL);
+            theDestTtl = (UInt16) ttlParser.ConsumeInteger(NULL);
         }
         // Each x=qt-relay-port line corresponds to one source stream.
         else if (sdpLine.NumEqualIgnoreCase(sRelayPort.Ptr, sRelayPort.Len))
@@ -176,7 +177,7 @@ void    RelaySDPSourceInfo::Parse(StrPtrLen* inSDPData)
             relayAddrParser.ConsumeUntil(NULL, StringParser::sDigitMask);
         
             // The first port is the source port for this stream
-            fStreamArray[theStreamIndex].fPort = relayAddrParser.ConsumeInteger(NULL);
+            fStreamArray[theStreamIndex].fPort = (UInt16) relayAddrParser.ConsumeInteger(NULL);
             if (fStreamArray[theStreamIndex].fPort & 1)
                 continue; //we only care about RTP ports
             
@@ -189,7 +190,7 @@ void    RelaySDPSourceInfo::Parse(StrPtrLen* inSDPData)
             for (UInt32 x = 0; x < fNumOutputs; x++)
             {
                 relayAddrParser.ConsumeWhitespace();
-                fOutputArray[x].fPortArray[theStreamIndex] = relayAddrParser.ConsumeInteger(NULL);
+                fOutputArray[x].fPortArray[theStreamIndex] = (UInt16) relayAddrParser.ConsumeInteger(NULL);
             }
             
             theStreamIndex++;
